feat: add profile preference to staff overview
